Improved HTS voice performance and reduced disk i/o.
This commit improves performance with HTS voices in two ways:
a) Instead of loading all HTS voice models on each utterance,
it only reloads the models if the voice name has changed (i.e. a
different voice is being used).

b) Disk usage is reduced: Before, speech parameters (mel cepstrum and
logf0 coefficients) were written to a temporal file and removed. Also, HTS text
features were passed to the HTS_Synthesize_Utt function by using a text file,
and most importantly the generated wavefile was saved to a temporal file and
reloaded from that file in an unnecessary intermediate step.

Now, by default speech parameters are not saved and deleted, HTS text features
are passed using a string, and the generated wavefile is loaded using a
stringstream instead of a file.

This commit requires the tokenstream_istream.diff patch in speech-tools.